Consolidating files in read
